fixup "Remove blocking calls from root and parent RPCs" (reduce session metadata)
This patch allows noux_tool_chain_auto to successfully start in okl4.
Without it, core tries to expand init's core-local heap partition in the
middle of a session request by init, which exhausts init's preserved
RAM. The underlying issue is that the session meta data is currently not
paid by the session quota. This, however, would require a bigger change
of the connection intrastructure - to be done.