Do not merge "base-sel4 core: try to flush page entry selectors as well"
On explicit user actions like RM unbind_thread or detach the selectors are already
of the no longer needed pagetable entries are already freed. But this is not done
when the pager thread in core resolves a user pagefault and runs into a full mapping
cache. In this case, the mapping cache gets fully flushed but the selectors
remain allocated. This is why, after a sufficient amount of pagefaults, core gets
an uncaught Bit_allocator::Out_of_indices that originates from the selector
allocation at the beginning of Vm_space::_map_page.

This occurs for example in the fb_bench test on x86_32 when the second test
starts.

This commit tries to clean up the selector and the pageframe copy of each
pagetable entry that is flushed. Unfortunately, in the x86_32 fb_bench this
leads to warnings

Warning: invalid signal-context capability
Warning: page-fault, pager_object: pd='init -> test-fb_bench' thread='ep' ip=0x5b1df pf-addr=0x401ff008

after or during the first mapping-cache flush (when the first test starts).

Ref #2268