Register: avoid warnings about oversized shifts
At some points, the framework caused compiler warnings about shifts that
xceed the argument type. E.g. if the width of an array item was equal to
the width of the access type of the array. This commit fixes this problem
for Register, Bitfield, and Register_array. But not for Bitset_* so far.

Fixes #2369