bit_allocator: add dynamic version
The normal Bit_allocator is a template that must be dimensioned at compile time.
For more dynamic allocations one might use an Allocator implementation instead.
However, usually with Allocator implementations the metadata costs for
allocations are hardly or impossible to forsee even if it is known what are
the allocations that will be made. This is the case for instance in a DHCP
server where the range of IP addresses is given through the configuration. The
metadata costs for the IP allocator can not be accounted to any client
sessions as they may not be existant on allocator construction but the servers
own RAM consumption shall not change after initialization which would be the
case if we'd use an Allocator here. Thus, what we want to use is the
new Bit_allocator_dynamic whose RAM consumption is static and forseeable and
whose backing store can be created dynamically.

Ref #2490