net: make packet print-functionality configurable
Provide a wrapper for all supported packet types that takes a packet and
a verbosity configuration. The wrapper object can than be used as
argument for a Genode log function and prints the packet's contents
according to the given configuration. The configuration is a distinct
class to enable the reuse of one instance for different packets.

There are currently 4 possible configurations for each protocol:
* NONE          (no output for this protocol)
* SHORT         (only the protocol name)
* COMPACT       (the most important information densely packed)
* COMPREHENSIVE (all header information of this protocol)

Ref #2490