WiP! Merge remove noux/net, uplift noux/minimal
Now that Berkley socket support for all libc-based components is
consolidated into the libc library, there is no need for two versions of
Noux. Noux applications now use the socket routines that are provided in
the libc library, and the 'socket' attribute of the 'libc' XML
configuration element propagates from the Noux configuration to each
child so that they may use socket_fs files instantiated by the Noux
parent.

Ref #2535