intel_fb.run: add acpica and poll
This solves problems with external monitors at laptops with docking
stations. Normally, when a docking station is plugged in or out, one of
the display interfaces of the laptop has to be switched between the
external ports of the laptop itself and those of the docking station.
This may be the responsibility of the ACPI depending on the hardware. It
is necessary to let the Intel FB poll for changes as there is yet not
communication protocol between ACPICA and Intel FB.

Fixes #2956