Refactor VFS errors
Remove VFS errors not returned by any of the VFS plugins and replace the
"Insufficient_buffer" exception with "WRITE_BLOCKED". This generates a
compile-time error when the VFS write call is used directly but the
"WRITE_BLOCKED" condition is not handled by a switch statement. Retrying
writes is not handled by the VFS, it is the callers responsibility to
dispatch I/O signals and call write again.

The "Insufficient_buffer" exception originated at the File_system client
within the VFS and was thrown when there was not enough space in the
packet stream buffer to fit an entire write operation. This condition is
now indicated by the "WRITE_BLOCKED" error, and the condition that a
write is larger than the entire packet stream buffer and will never
complete is indicated by a "WRITE_ERR_INVALID".

The "WRITE_BLOCKED" error is also propagated by plugins that implement
TCP to indicated send window congestion.

The VFS server has been updated to handle this new write error and the
packet handling logic is refactored more broadly.

Ref #2944
Fix #2954