heap: free DS on Out_of_ram during attach
Previously, we had only one 'try'/'catch' for 'alloc' and 'attach' of
new dataspaces. In the 'catch' we did not clean up but merely returned
zero. This could lead to leaking dataspaces that were successfully
allocated but failed to be attached. Now we free them if the 'attach'
throws an exception.