cbe_manager: GUI for controlling CBE devices
The current version of the CBE manager is not thought for productive use but
for mere demonstrational purpose! Please refrain from storing sensitive data
with it!

The CBE manager is a graphical front end for creating and controlling a virtual
encrypted block device using Genodes Consistent Block Encrypter (CBE). The
manager also takes care of creating and managing a Trust Anchor for securing
the block device and formatting the block device with the ext2 file system
which is then provided to clients through Genodes File System service.

The commit also adds a not fully automatic test run script for the manager and
a package for Sculpt.

For more details see the file _gems/src/app/cbe_manager/README_.

Fixes #4032