vfs lib: move implementations from header to unit
Debugging problems with the VFS or developing it in general is currently a
tedious work as many implementations of the VFS are located in public headers
that are included by lots of compile units. Thus, changing the implementations
triggers many re-compile processes. This commits starts changing this by
moving some of the implementations to dedicated compilation units.

Ref #4032