vfs/cbe_trust_anchor: sync hashfile-handle close
In order to synchronize the closing of the hash-file-handle in the CBE TA VFS
after a file access with the termination of that file access, we issue a "sync"
operation at that file after having started the file access. Otherwise, the
client of the CBE TA VFS might refer to the handle at a time were the CBE TA
VFS has already thrown it away resulting in all kind of problems.

Ref #4032