file_vault: truncate image to new size on resizing
When resizing, the CBE doesn't write to new leaf blocks, it only writes to
new inner-node blocks. As the cbe-image size only adapts when writing to a
file offset greater the file size and CBE block addresses are directly used
for the file offset, the image size only adapts now and then when the resizing
creates a new inner node. This would result in false prognosis of the future
image size in the resizing dialog. Thus, before expanding the client FS or the
snapshot buffer, we now truncate the file to the new size.

This commit renames the former 'new_empty_file' helper component to
'truncate_file' and removes the additional ftruncate(file, 0) from its body (so
the tool now issues only one ftruncate with the desired size).