file_vault: hide optional controls
This hides the snapshots menu, the private key and user passphrase menu and
renames the snapshot buffer journaling buffer in order to give the user a less
bloated and more coherent interface. This is because these menus have no
real function yet (e.g. snapshots can be created and discarded but not
accessed).