base-hw: prevent optimizing out irq_id in cpu.cc
If the 'irq_id' stack variable that is handed over to 'Pic::take_request' is
not used before handing it over, and the compiler doesn't know the
implementation of 'take_request' (as it's the case with base-spunky), the
compiler seems to re-order creating the variable to a point after the
'take_request' call. In case of base-spunky, this led to memory corruption. We
fix the issue by initializing the variable before calling 'take_request'.