vfs/fatfs: remove uneffective attributes
Apparently, the former XML attributes to the plugin 'drive' and 'codepage' had
no effect. I tested them in a scenario with the VFS block server on a
disk-image boot-module as back end. Regardless of the 'drive' value, the block
session label was always "0". Regardless of the 'codepage' value, the FAT on
the disk image succeeded to mount when not using '--codepage' for 'mkfs.fat'
and failed to mount when using '--codepage' to specify a supported but foreign
codepage for 'mkfs.fat' (e.g. "720").

Ref #4220