nic_router: replace term "uplink" with "NIC client"
! Attention: Changes configuration interface of the NIC router in a way that
  may break scenarios that use the component without proper adjustment !

The term "uplink" for network interfaces in the router that have a NIC session
client as back end was introduced in a time when "Uplink" sessions didn't exist
yet. Now, they do and, although both an "uplink" and an "Uplink" session
normally describe a network session towards a network device driver, they are
based on two different service types. This can easily cause confusion when
integrating the router (the <uplink> is not related to "Uplink" sessions) or
trying to understand its functioning (an 'Uplink' object has nothing to do
with the "Uplink" service).

Therefore, this commit introduces the more specific term "NIC client" for an
interface that is based on a NIC session requested by the router. This doesn't
imply any semantic changes at the NIC router. However, the commit also brings a
broader update of the router's README and removes the term "downlink" that was
used only in documentation to refer to interfaces backed by a NIC session
provided by the router. The term "downlink" was only associated with this
meaning because it is the natural counterpart to "uplink". This isn't
appropriate anymore as the terms for interface back-ends have moved to a more
technical level.

How to adjust scenarios that use the NIC router properly to this commit:
At each occurrence of the '<uplink ...>' tag in a NIC router configuration
replace the tag name 'uplink' with 'nic-client'. The rest of the tag stays the
same.