util/register: registers on unaligned integers
The register framework so far didn't support writing bitfields in registers
on unaligned integers. This commit introduces two new methods
'Genode::Register::Bitfield::set_unaligned' and
'Genode::Register::Bitfield::clear_unaligned' for that. The new methods also
use a reference to the register's access-integer type but with alignment 1.

The commit also extends the 'mmio' test in order to test the new feature.

Fixes #4245