QUICKFIX "cbe_trust_anchor: skip closing invalid handles"
Since commit "vfs/cbe_trust_anchor: close handles correctly", the CBE trust
anchor lib sporadically runs into a nullpointer when wanting to close a VFS
handle. This commit quickfixes it by adding a check-and-return for nullpointers
to the '_close_handle' function. However, in the long run, I will try to find
and fix the reason for the nullpointer instead.