lx_emul & wireguard & wifi: centralized random.cc
Both the Wifi driver and the WireGuard port used local implementations for
their source of randomness. Wifi used a Xoroshiro sponge for rapid generation
of random values but initialized this sponge always with the same static seed
value. WireGuard requested each random byte directly from jitterentropy, which
is considered to be very time intensive.

This commit removes the local variants of random.cc and introduces a new
centralized lx_emul/random.cc . The new variant combines the former approaches,
so, that jitterentropy is accessed only in order to generate a random seed for
a Xoroshiro sponge. Front-end requests for random values are then fulfilled
efficiently via the sponge.

The Xoroshiro sponge is currently seeded only once the first time it gets
used (no re-seeding).

Ref #4397