lx_block: support special block devices
With the SmartCard-HSM 4K Swissbit MicroSD we observed that lx_block
failed to drive a special block device correctly.

With this commit lx_block assumes devices with size 0 ('stat' call) to be
special block devices. The treatment of special block devices differs from
that of normal block devices as follows:

* The device file gets opened with O_DIRECT and O_SYNC.
* The block count and block size are determined via 'ioctl' calls.
* It is assumed that the block size is in fact its sector size if the sector
  size is smaller. In linux, the block size changes to the sector size in this
  case when mounting the device.
* Read and write requests are executed only if the provided buffer is block
  size aligned. Otherwise a warning is printed and an IO error is thrown.