base-hw scheduler: print method
Prints the internal state of the scheduler as JSON text to the given output.
Adds a new constructor parameter and member "ID" to the CPU share class that
are needed only for easier reference of the shares in the printed output. The
ID is set to 0 by default for backwards compatibility.

Ref #4151