base-hw scheduler test: fix wrong assumptions
The test used to check whether the act of setting a share ready outdates the
head or not. However, with the most recent version of the scheduler, this
check is not possible anymore. We can merely check whether the head is
outdated after setting the share ready. This commit adapts the expectations of
the test to the new semantics of the check.

Ref #4151