base-hw ipc: methods whether channels are blocking
This introduces the notion of an in and an out channel inside the IPC node.
The plan is, that the whole IPC node is split into two state machines (in and
out, i.e. receiving end and sending end) that work mostly independent from each
other in order to simplify the implementation. In this first step, two private
methods _in_waiting and _out_sending are introduced that return whether the
receiving end is blocking for incoming requests respectively whether the
sending end is blocking for an incoming reply. These methods still work with
the existing state values, so the commit does not change any semantics.