base-hw ipc: fix "cancel waiting" method
The Ipc_node::_cancel_waiting method used to check whether the in _or_ the out
channel was waiting and unblocked only one of the two. This is fine considering
the still active state where IPC node has only one global state and in and out
can never be blocking at the same time. However, as we want to separate the
states of in and out channel more and more, we can now ignore this background
knowledge and handle both channels independently in Ipc_node::_cancel_waiting.