base-hw ipc: introduce state "destruct"
During destruction of an IPC node, we used to leave the node in its former
state. However, this might cause some unnecessary overhead as the generic code,
while dissolving the node, must assume that the node is still operational.
The new state allows for a distinctive handling of the destruction phase.