base-hw ipc: use "cancel send" in destructor
This replaces the former individual implementation of cancelling the relation
of a sender towards an IPC node in destruction with a call to the already
existing method Ipc_node::_cancel_send. Although the implementations differ,
it is a valid assumption that Ipc_node::_cancel_send must cover this use
case as well, which it actually does.