base-hw ipc: get rid of "caller" member variable
Instead of having an extra reference to the sending node whose request we are
currently processing (besides the existing in queue), this node can simply stay
the head of the in queue until it receives a reply from us. This idea can be
implemented with some small adaptions only to the rest of the state handling.