nic_router: no denied sessions with dynamic config
The NIC router used to announce its services directly on startup. This has a
negative side effect in scenarios where the NIC-router configuration is
dynamically generated and therefore potentially empty at the beginning. If a
session request of a client reaches the NIC router before the configuration
has turned non-empty and valid, the request is denied. It is common among
network client components such as NIC drivers to treat this as a fatal error
and terminate. This can be prevented by deferring the announcement of the NIC
router services to the point in time where the NIC router receives its first
valid configuration, which is just what this commit is doing.

Fix #4738