sculpt network: wired without router interception
Adds a new button "Wired (No Router)" to the network dialog in setups that
support wired NIC. Similar to the "Wired" option, this button spawns a NIC
driver, but in contrast to "Wired" it accompanies it with a NIC-Uplink
component as server instead of a NIC router. This enables raw access to the
network connected to the NIC without any packet modifications. It therefore
provides a way of circumventing the current protocol limitations of the
Genode-native network stack at the cost of less protection.