tresor: fix generation override of user snapshot
When a user requested "create snapshot" twice without intermediate changes to
the virtual block device, the Tresor library created the first user snapshot
and then used to simply raise the generation of this first one on the second
call. This caused the generation of once created user snapshots to change
without user notification.

With this commit, the Tresor library will detect the situation and simply
return the unaltered generation of the first user snapshot on the second
call.