tresor_tester: replace list- with check-snapshots
Removes the <list-snapshots> command from the Tresor tester and adds the new
<check-snapshots> command as replacement. The new command checks whether all
user snapshots known to the Tresor tester are also known to the Tresor
container and vice versa.

The Tresor tester keeps track of the active user snapshots via its command list
and the snapshot IDs provided in it. After a successful "create snapshot", the
tester adds a reference to the generation of the new snapshot. After a
successful "discard snapshot", the tester removes all references to the
generation of the removed snapshot (multiple snapshot IDs can reference the
same generation). Before an "initialize", the tester removes all references.