tresor: discard snapshots only when writing the sb
The Tresor lib used to discard disposable snapshots on the beginning of each
Supeblock-Control request. This was motivated by the performance overhead that
is caused by keeping snapshots. However, removing a snapshot from the in-RAM
superblock without updating the last secured on-disc superblock is a bad idea.
In this constellation the Tresor lib might mark physical blocks that still form
part of the last-secured-gen snapshot in the on-disc superblock because it
doesn't find a snapshot that references it anymore.

That said, it seems best to move the snapshot "garbage collection" to the
beginning of the state machine that secures the superblock. If the securing
fails after removing the snap from the in-RAM SB, the latter should in any
case not beused anymore. Furthermore, this shouldn't affect performance
significantly as securing the SB is done on each request except common VBA
access.