nic_router: keep links on resource exhaustion
This removes the garbage collection that the router used to do when a session
ran out of quota. This is motivated by the fact that the garbage collection was
rather simple and removed connection states regardless of their current state.

This caused problems, e.g., when the router dropped an established TCP
connection with NAPT and on the next matching packet re-created it with
different NAPT port, thereby breaking the connection. With this commit,
existing connections are prioritized over new ones during resource exhaustion
and the packets that attempt to create a new connection in such a state are
dropped with a warning in the log (verbose_packet_drop="yes").

Note that the state resolves itself with time as existing connections time out
or are closed by peers.

Ref #4729