xml_node: support attribute access via lambda
Similar to how XML nodes can be accessed via a lambda, this commit implements
accessing attributes of XML nodes via a lambda. There are two new methods in
Xml_node: 'with_optional_attribute' calls a functor only if the attribute was
found and 'with_attribute' additionally calls a functor if the attribute could
not be found. Both methods are const and call the "found" functor with one
argument 'Xml_attribute const &'.

Ref #4729