nic_router: smarter emergency free on exhaustion
Re-implements an emergency freeing of resources on exhaustion of session quota.
In contrast to the past one, the new algorithm is executed directly where the
exhaustion occurs. Instead of interupting the packet handling and restart it
from the beginning after the freeing action, packet handling is now continued
at the point of exhaustion (if enough resources could be freed).  Furthermore,
the new algorithm frees only as much RAM as needed instead of thowing away a
fixed number of objects, like in the past. And finally, the router now drops
ICMP first, then UDP, then TCP (better reflects priorities) and refrains from
dropping TCP connections in the ESTABLISHED state. If the router cannot free a
sufficient amount of resources, the packet that caused the exhaustion is
dropped with a warning (verbose_packet_drop="yes").

Ref #4729