nic_router_flood.run: no verbose packet drop
This test used to have the verbose packet drop at the NIC router enabled
because the router, when in a session-quota emergency, used to prioritize new
L4 connections over existing ones and was therefore expected to not drop
a single packet in this test. However, the emergency policy has changed with
the latest router updates to prioritizing established TCP connections over new
ones, which caused a lot of verbose output and rendered the test to slow on at
least x86_64 foc.